feat: port read-side from DB procs to .NET + error-handling foundation
Reimplements business logic from PostgreSQL fn_*/usp_*/vw_* routines in C#
(goal: move logic out of the DB), verified against the live procs as a parity
oracle. Covers the entire read side; write-side validation (P4) is next.

Endpoints (parity-verified vs procs/views):
- GET /api/customer (paginated, status, search) + /{id}   [fn_get_customers, fn_get_customer_by_id]
- GET /api/booking (paginated + 8 filters)                [usp_get_bookings]
- GET /api/booking/Active (tier/usedMinutes/payments/logs) [vw_activebookings]
- GET /api/child/search-unified                           [fn_search_unified, 4-tier]
- GET /api/purchase, /customer, /purchase-bookings-by-id  [vw_purchases, fn_get_purchase_by_customer_id, usp_get_purchase_bookings_by_id]
- GET /api/table?name, GET /api/dashboard/summary         [vw_tables, 6x fn_get_dashboard_*]

Foundation:
- ProblemDetails (RFC7807) via IExceptionHandler + BusinessRuleException/NotFoundException
- ApiResponse<T> envelope across all 11 controllers (replaces CustomResponse)
- camelCase JSON policy; CustomerController [Authorize] enabled; register no longer leaks password
- SubscriptionService: shared billing core (used-minutes, best package, subscriber tier)

Correctness fixes surfaced by parity testing (old behavior was buggy; not carried over):
- money/hours int -> decimal (int truncated 1.2 -> 1, corrupting billing)
- usedHours was actually minutes; old .NET wrongly /60 -> usedMinutes
- top_customers summed only Completed bookings (proc sums all statuses)
- TZ: connection Timezone=Europe/Istanbul + TR today (dashboard date/hour buckets)
- Table entity had phantom createdat/updatedat columns; BookingLogType.Edit was missing
- usp_get_purchase_bookings_by_id referenced non-existent b.customerId (proc errors) -> derived from child.parentId
- net9 runtime absent on host -> RollForward=Major

Also: .gitignore now excludes Claude planning docs (CLAUDE/MIGRATION-PLAN/PARITY/REFACTORING .md).

+++ Models/DTOs/Booking/BookingCreateDto.cs
@@ -10,7 +10,7 @@ namespace Iroh.Models.DTOs.Booking
         public DateTime? startTime { get; set; }
         public DateTime? endTime { get; set; }
         public required BookingStatus status { get; set; }
-        public int? price { get; set; }
+        public decimal? price { get; set; }
         public int? childId { get; set; }
         public string? note { get; set; }
     }
