+++ Models/DTOs/PurchasePayment/PurchasePaymentCreateDto.cs
@@ -6,10 +6,10 @@ namespace Iroh.Models.DTOs.PurchasePayment
     public class PurchasePaymentCreateDto
     {
         [Required]
-        public required int hours { get; set; }
+        public required decimal hours { get; set; }
 
         [Required]
-        public required int price { get; set; }
+        public required decimal price { get; set; }
 
         [Required]
         public required int purchaseId { get; set; }
